refactor: execute runs in separate task; remove is_a_job
